- Made content editable.  - Added condition to show edit option if my post or visitor or anonymous is set to true.  - Added API call to update post & reset old value on ‘cancel’ click  - Displayed ‘Edited’ icon if text is edited  - Added popover message when click on ‘Edited’ icon  - Disabled message click if message is from bot  - Made different card for announcement and display data according to it.  - Used announcements to display image on all cards.

+++ src/screens/ChatScreen.js
@@ -582,16 +582,21 @@ export default function ChatScreen(props) {
           : 'Publish',
         onPress: () => approveItem(selectedMessageClone),
       },
-      {
-        title: 'Delete',
-        onPress: () => deleteItemAlert(selectedMessageClone),
-      },
     ];
+    const amMod = reduxState.selectedEvent.moderators.includes(
+      reduxState.user.accountId,
+    );
     if (
       selectedMessageClone.author.id === reduxState.user.accountId ||
-      selectedMessageClone.visitor ||
-      selectedMessageClone.anonymous
+      (amMod &&
+        (['Q', 'M', 'A', 'C'].includes(selectedMessageClone.type) ||
+          selectedMessageClone.visitor ||
+          selectedMessageClone.anonymous))
     ) {
+      options.push({
+        title: 'Delete',
+        onPress: () => deleteItemAlert(selectedMessageClone),
+      });
       options.push({
         title: 'Edit',
         onPress: () => editItemPress(selectedMessageClone),
